% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_OmniPeak}
\alias{run_OmniPeak}
\title{Run the OmniPeak Application}
\usage{
run_OmniPeak()
}
\description{
Run the OmniPeak Application locally. Shiny deployment
\href{https://plyush1993.shinyapps.io/OmniPeak/}{here}.
}
